add create label option to label assignment modal
