+++ apple/OmnivoreKit/Sources/App/Views/Labels/ApplyLabelsView.swift
@@ -1,50 +1,7 @@
-import Combine
 import Models
 import Services
 import SwiftUI
 import Views
-
-final class ApplyLabelsViewModel: ObservableObject {
-  private var hasLoadedInitialLabels = false
-  @Published var isLoading = true
-  @Published var selectedLabels = [FeedItemLabel]()
-  @Published var unselectedLabels = [FeedItemLabel]()
-
-  var subscriptions = Set<AnyCancellable>()
-
-  func load(item: FeedItem, dataService: DataService) {
-    guard !hasLoadedInitialLabels else { return }
-
-    dataService.labelsPublisher().sink(
-      receiveCompletion: { _ in },
-      receiveValue: { [weak self] allLabels in
-        self?.isLoading = false
-        self?.hasLoadedInitialLabels = true
-        self?.selectedLabels = item.labels
-        self?.unselectedLabels = allLabels.filter { !item.labels.contains($0) }
-      }
-    )
-    .store(in: &subscriptions)
-  }
-
-  func saveChanges(itemID: String, dataService: DataService, onComplete: @escaping ([FeedItemLabel]) -> Void) {
-    dataService.updateArticleLabelsPublisher(itemID: itemID, labelIDs: selectedLabels.map(\.id)).sink(
-      receiveCompletion: { _ in },
-      receiveValue: { onComplete($0) }
-    )
-    .store(in: &subscriptions)
-  }
-
-  func addLabel(_ label: FeedItemLabel) {
-    selectedLabels.insert(label, at: 0)
-    unselectedLabels.removeAll { $0.id == label.id }
-  }
-
-  func removeLabel(_ label: FeedItemLabel) {
-    unselectedLabels.insert(label, at: 0)
-    selectedLabels.removeAll { $0.id == label.id }
-  }
-}
 
 struct ApplyLabelsView: View {
   let item: FeedItem
@@ -52,7 +9,7 @@ struct ApplyLabelsView: View {
 
   @EnvironmentObject var dataService: DataService
   @Environment(\.presentationMode) private var presentationMode
-  @StateObject var viewModel = ApplyLabelsViewModel()
+  @StateObject var viewModel = LabelsViewModel()
 
   var body: some View {
     NavigationView {
@@ -95,6 +52,19 @@ struct ApplyLabelsView: View {
               }
             }
           }
+          Section {
+            Button(
+              action: { viewModel.showCreateEmailModal = true },
+              label: {
+                HStack {
+                  Image(systemName: "plus.circle.fill").foregroundColor(.green)
+                  Text("Create a new Label")
+                  Spacer()
+                }
+              }
+            )
+            .disabled(viewModel.isLoading)
+          }
         }
         .navigationTitle("Assign Labels")
         .navigationBarTitleDisplayMode(.inline)
@@ -116,6 +86,9 @@ struct ApplyLabelsView: View {
               label: { Text("Save") }
             )
           }
+        }
+        .sheet(isPresented: $viewModel.showCreateEmailModal) {
+          CreateLabelView(viewModel: viewModel)
         }
       }
     }
